requirements.txt
anthropic>=0.50.0
diffusers>=0.33.1
einops>=0.6.0
fastapi>=0.95.0
groq>=0.23.1
huggingface-hub>=0.30.2
jinja2>=3.1.2
numpy>=1.24.0
openai>=1.76.0
onnxruntime[gpu]>=1.21.1
pillow>=9.4.0
python-multipart>=0.0.6
sageattention
torch>=2.6.0
torchvision>=0.21.0
transformers>=4.46.2
uvicorn>=0.22.0
watchdog
xformers>=0.0.29.post3
# Additional dependencies for Wan module
easydict
imageio
imageio-ffmpeg
ftfy
opencv-python>=4.9.0.80
--extra-index-url https://download.pytorch.org/whl/cu126